refactor(serializers): move CleanTextMixin to lib/serializers/mixins
Move CleanTextMixin from lib/validators.py to lib/serializers/mixins.py
to align with the existing DAB pattern where serializer mixin classes
live alongside ImmutableFieldsMixin and EmailAdminOnlyMixin.

The raw validator primitives (resource_name_validator,
DANGEROUS_PATTERNS, DEFAULT_NAME_FIELDS) remain in lib/validators.py
where they belong.

ANSTRAT-1756
